Code a fib seq generator position finder
This one was more intricate and I ended up
going to SO to find a solution to it.  I got the
original code down by looking up a fibonacci seq
definition and it worked fine.  The refactor
with `fill()` and `reduce()` ended up taking some
time to understand and parse through.  I'll likely
end up having to look this solution up again if
I end up having to generate a sequence and pull a position
from it.  I did end up having to tweak the code from
the solution I found on SO to fit my particular needs.